events: Move msgtype types in separate files
